Seperated logger into multiple files
