+++ Sources/HydrogenReporter/LogComplexity.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+// MARK: Logger Complexity
+public enum LogComplexity: String, Codable {
+    case simple
+    case complex
+}
